Implement a single test explicitly for flocker-deploy.
