+++ flocker/acceptance/endtoend/__init__.py
@@ -1,0 +1,7 @@
+# Copyright ClusterHQ Inc.  See LICENSE file for details.
+
+"""
+Tests for the end-to-end internal operation of the Flocker cluster.
+
+In particular this does not cover testing interactions with external systems.
+"""
